requirements.txt
accelerate==0.25.0
certifi==2023.11.17
charset-normalizer==3.3.2
cmake==3.28.1
contourpy==1.1.1
cycler==0.12.1
decord==0.6.0
diffusers==0.19.3
docopt==0.6.2
einops==0.7.0
ffmpeg==1.4
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
huggingface-hub==0.20.1
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
lit==17.0.6
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pipreqs==0.4.13
psutil==5.9.7
pyparsing==3.1.1
pyre-extensions==0.0.29
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scipy==1.10.1
segment-anything==1.0
six==1.16.0
sympy==1.12
tokenizers==0.15.0
torch==2.0.1
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.66.1
transformers==4.36.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
xformers==0.0.20
yarg==0.1.9
zipp==3.17.0
